Set schema at repo level
