+++ packages/cli/src/databases/repositories/schema.ts
@@ -1,0 +1,10 @@
+/**
+ *
+ */
+export class Schema {
+	constructor(private data: unknown = {}) {}
+
+	getFieldNames(): string[] {
+		return [];
+	}
+}
